// src/store/auth/mutation-types.ts
export enum AuthMutationTypes {
  SET_IS_LOGGED_IN = 'AUTH/SET_IS_LOGGED_IN',
  SET_USER = 'AUTH/SET_USER',
  SET_SIGN_UP_ERROR = 'AUTH/SET_SIGN_UP_ERROR',
  SET_LOGGED_IN_ERROR = 'AUTH/SET_LOGGED_IN_ERROR',
  RESET_USER = 'AUTH/RESET_USER',
  SET_IS_LOADING = 'AUTH/SET_IS_LOADING',
  SET_IS_LOADED = 'AUTH/SET_IS_LOADED',
  SET_SIGN_UP_FULLFILLED = 'AUTH/SET_SIGN_UP_FULLFILLED'
}